feat: integrate GitHub star count into Showcase layout and components
